cloud_google/CloudFunction/HTTPTriggerPython/requirements.txt
# Function dependencies for python modelue
google-cloud-bigquery
